% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris.R
\docType{data}
\name{iris}
\alias{iris}
\title{Measurements of Iris species}
\format{
A data frame with 150 rows and 5 variables:
\describe{
\item{Sepal.Length}{length of sepal, in cm}
\item{Sepal.Width}{width of sepal, in cm}
\item{Petal.Length}{length of petal, in cm}
\item{Petal.Width}{weight of the diamond, in carats}
\item{Species}{Species of iris, factor with three levels}
...
}
}
\source{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.
}
\usage{
iris
}
\description{
A data introduced by Ronald Fisher in his 1936 paper: The use of multiple measurements in taxonomic problems.
Widely used in R. All measurements in centimeters
}
\keyword{datasets}
