% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantGrowth.R
\docType{data}
\name{PlantGrowth}
\alias{PlantGrowth}
\title{Results of a Plant Growth Experiment}
\format{
A data frame with 30 rows and 2 variables:
\describe{
\item{weight}{weight of dried plant growth, numerical}
\item{group}{group assignment, categorical with three levels}
...
}
}
\source{
Dobson, A. J. (1983) An Introduction to Statistical Modelling. London: Chapman and Hall.
}
\usage{
PlantGrowth
}
\description{
Comparison of dried weight of plants from an experiment comparing a control group to two different treatment conditions.
}
\keyword{datasets}
