% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable.R
\name{SummaryTable}
\alias{SummaryTable}
\title{PHSMChart
This function will build a chart showing the epicurve against PHSM measures}
\usage{
SummaryTable(dataset)
}
\arguments{
\item{dataset}{qry_COVID_cases_by_date_final.csv}
}
\description{
PHSMChart
This function will build a chart showing the epicurve against PHSM measures
}
