% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_freq_perc.R
\name{table_helper}
\alias{table_helper}
\title{Table Helper}
\usage{
table_helper(df, var)
}
\arguments{
\item{df}{is the data frame to create summaries from}

\item{var}{is the non-numeric variable from data frame df}
}
\value{
Tibble containing the number of unique observations in each group and the percentage of the total observations that the group contains
}
\description{
This function helps take a data frame and a chosen variable and counts the frequency of
observations within each group then computes what percentage those observations are of the total observations of that variable.
}
