% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{test.data}
\alias{test.data}
\title{Dataset to test count_freq_perc() funciton}
\format{
A data frame with 5 rows and 5 variables:
\describe{
\item{ID}{list of ID, numerical, no NA}
\item{colour}{colour categories, no NA, four levels}
\item{sex}{sex categories, includes NA, three levels}
\item{age}{made up age, numerical, includes NA}
\item{nothing}{vector completely NA}
...
}
}
\source{
created by author
}
\usage{
test.data
}
\description{
a created test data set to have special test cases and expected data to test this funciton.
}
\keyword{datasets}
