% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stats_table.R
\name{tidy_stats_table}
\alias{tidy_stats_table}
\title{Generate table with summary statistics}
\usage{
tidy_stats_table(stats_table, var)
}
\arguments{
\item{dat}{stats table output from `var_summarise()`}
}
\value{
Stats output field_label tidydied
}
\description{
Generate table with summary statistics
}
\examples{
SummaryTable::var_summarise(iris, var = "Petal.Length", treatment = "Species") \%>\% tidy_stats_table()

}
