% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{var_summarise}
\alias{var_summarise}
\title{Generate table with summary statistics}
\usage{
var_summarise(dat, var = NULL, treatment = NULL, nonparametric = NULL)
}
\arguments{
\item{dat}{data.frame with data to summarise}

\item{var}{variable or list of variables to summarise}

\item{treatment}{variable to stratify summary variables by}

\item{nonparametric}{bolean to indicate if median and IQR should be calculated instead of mean}
}
\value{
A tibble with data summary statistics is returned.
}
\description{
Generate table with summary statistics
}
\examples{
SummaryTable::var_summarise(iris, var = "Petal.Length", treatment = "Species")
}
