% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{Summary_function}
\alias{Summary_function}
\title{Group Summary Statistics}
\usage{
Summary_function(data, cat, y)
}
\arguments{
\item{data}{\itemize{
\item refers to the dataset you want to create the summary statistics for.
}}

\item{cat}{\itemize{
\item refers to the category that you wish to group by.
}}

\item{y}{\itemize{
\item refers to the numerical variable that you wish to summarise, in relation to the groups defined in cat.
}}
}
\value{
A summary table providing the  minimum, maximum, average (mean), range and standard deviation of variable y as grouped by categorical variable cat.
}
\description{
This function will provide summary statistics, for a numerical variable y when grouped by another variable x.
The statistics generated are; minimum, maximum, average (mean), range and standard deviation.
}
\examples{
Summary_function(gapminder::gapminder, country, lifeExp)
Summary_function(flow_sample, extreme_type, flow)
}
