% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{Summary_function}
\alias{Summary_function}
\title{Group Summary Statistics}
\usage{
Summary_function(data, cat, y)
}
\value{
A summary table providing the  minimum, maximum, average (mean), range and standard deviation of variable y as grouped by x.
}
\description{
This function will provide summary statistics, for a numerical variable y when grouped by another variable x.
The statistics generated are; minimum, maximum, average (mean), range and standard deviation.
}
\examples{
Summary_function(gapminder, country, lifeExp)
Summary_function(flow_sample, extreme_type, flow)
}
