% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable.R
\name{SummaryTable}
\alias{SummaryTable}
\title{SummaryTable
Function giving the summary table}
\usage{
SummaryTable(dataset)
}
\arguments{
\item{dataset}{qry_COVID_cases_by_date_final.csv}
}
\description{
SummaryTable
Function giving the summary table
}
